fix arrumando comentarios

+++ api-agendamento-serviços/src/main/java/com/automotiva/estetica/rick/api_agendamento_servicos/dto/CategoriaDashboardDto.java
@@ -3,6 +3,7 @@ package com.automotiva.estetica.rick.api_agendamento_servicos.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,6 +11,6 @@ import java.util.List;
 
 public class CategoriaDashboardDto {
     private String categoria;
-    private Double totalCategoria;
+    private BigDecimal totalCategoria;
     private List<ServicoDashboardDto> servicos;
 }
