+++ api-agendamento-serviços/src/main/java/com/automotiva/estetica/rick/api_agendamento_servicos/dto/ServicoDashboardDto.java
@@ -3,11 +3,13 @@ package com.automotiva.estetica.rick.api_agendamento_servicos.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ServicoDashboardDto {
     private String servico;
-    private Double faturamento;
+    private BigDecimal faturamento;
 //    private Double lucro;
 //    private Double percentual;
 }
